+++ crates/bin/sierra-compile/Cargo.toml
@@ -11,7 +11,8 @@ anyhow.workspace = true
 log.workspace = true
 clap.workspace = true
 indoc.workspace = true
-serde_json = "1.0.116"
+serde.workspace = true
+serde_json.workspace = true
 
 cairo-lang-sierra = { path = "../../cairo-lang-sierra", version = "~2.6.3" }
 cairo-lang-sierra-to-casm = { path = "../../cairo-lang-sierra-to-casm", version = "~2.6.3" }
