+++ crates/cairo-lang-sierra-to-casm/Cargo.toml
@@ -19,7 +19,7 @@ indoc.workspace = true
 itertools = { workspace = true, default-features = true }
 num-bigint = { workspace = true, default-features = true }
 num-traits = { workspace = true, default-features = true }
-serde = { version = "1.0.200", default-features = false, features = ["derive"] }
+serde = { workspace = true, default-features = true, features = ["derive"] }
 thiserror.workspace = true
 
 [dev-dependencies]
