+++ crates/cairo-lang-sierra-to-casm/src/compiler.rs
@@ -1,4 +1,3 @@
-use serde::ser::Error;
 use serde::Serialize;
 use std::fmt::Display;
 
@@ -174,60 +173,6 @@ impl CairoProgram {
             bytecode.extend(instruction.assemble().encode().into_iter())
         }
         AssembledCairoProgram { bytecode, hints }
-    }
-}
-
-pub struct CairoProgramWithSierraContext<'a> {
-    pub cairo_program: &'a CairoProgram,
-    pub sierra_program: &'a Program,
-}
-
-impl<'a> CairoProgramWithSierraContext<'a> {
-    pub fn new(
-        cairo_program: &'a CairoProgram,
-        sierra_program: &'a Program,
-    ) -> CairoProgramWithSierraContext<'a> {
-        CairoProgramWithSierraContext { cairo_program, sierra_program }
-    }
-}
-
-impl<'a> Serialize for CairoProgramWithSierraContext<'a> {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        use serde::ser::SerializeMap;
-        let mut map = serializer.serialize_map(None)?;
-        let main_func = self
-            .sierra_program
-            .find_function("::main")
-            .ok_or_else(|| S::Error::custom("Main function not found"))?;
-        let entry_point = main_func.entry_point.0;
-        let builtins = vec![
-            "pedersen_builtin",
-            "range_check_builtin",
-            "bitwise_builtin",
-            "ec_op_builtin",
-            "poseidon_builtin",
-        ];
-        let assembled_cairo_program = self.cairo_program.assemble();
-        map.serialize_entry("bytecode", &assembled_cairo_program.bytecode)?;
-        map.serialize_entry("hints", &assembled_cairo_program.hints)?;
-        map.serialize_entry("entry_point", &entry_point)?;
-        map.serialize_entry("builtins", &builtins)?;
-        let debug_info = self.cairo_program.debug_info
-            .sierra_statement_info
-            .iter()
-            .map(|statement_debug_info| {
-                (
-                    statement_debug_info.start_offset,
-                    statement_debug_info.instruction_idx,
-                )
-            })
-            .collect::<Vec<(usize, usize)>>();
-        map.serialize_entry("debug_info", &debug_info)?;
-        map.serialize_entry("consts_info", &self.cairo_program.consts_info)?;
-        map.end()
     }
 }
