+++ crates/cairo-lang-utils/Cargo.toml
@@ -12,9 +12,9 @@ itertools = { workspace = true, features = ["use_alloc"] }
 num-bigint.workspace = true
 num-traits.workspace = true
 hashbrown = { workspace = true, features = ["serde"] }
-serde = { workspace = true, features = ["alloc", "derive"], optional = true }
 
 # Optional
+serde = { workspace = true, features = ["alloc", "derive"], optional = true }
 schemars = { workspace = true, features = ["preserve_order"], optional = true }
 env_logger = { workspace = true, optional = true }
 time = { workspace = true, optional = true }
@@ -33,7 +33,7 @@ std = [
   "indexmap/std",
   "num-bigint/std",
   "num-traits/std",
-  "serde/std"
+  "serde?/std"
 ]
 serde = ["dep:serde", "num-bigint/serde", "indexmap/serde"]
 schemars = ["std", "serde", "dep:schemars"]
